Refactored interfaces into new files, added a path singleton
